@@ -1,14 +1,13 @@
 {
 	"name": "react-wp-posts-fetcher",
-	"version": "0.1.0",
+	"version": "2.0.0",
 	"homepage": "https://asikur22.github.io/react-wp-posts-fetcher/",
 	"dependencies": {
-		"axios": "^0.19.0",
+		"axios": "^1.8.4",
 		"is-url": "^1.2.4",
-		"react": "^16.8.6",
-		"react-dom": "^16.8.6",
-		"react-scripts": "^3.0.1",
-		"react-select": "^2.4.2"
+		"react": "^19.0.0",
+		"react-dom": "^19.0.0",
+		"react-scripts": "^5.0.1"
 	},
 	"scripts": {
 		"predeploy": "npm run build",
@@ -23,6 +22,6 @@
 	},
 	"browserslist": [ ">0.2%", "not dead", "not ie <= 11", "not op_mini all" ],
 	"devDependencies": {
-		"gh-pages": "^2.0.1"
+		"gh-pages": "^6.3.0"
 	}
 }
