@@ -3,7 +3,7 @@ import React, { Component } from 'react';
 export class Header extends Component {
 	render() {
 		return (
-			<div style={{ backgroundColor: '#f4f4f4' }}>
+			<div className='py-4' style={{ backgroundColor: '#f4f4f4' }}>
 				<h1>WP Posts Feacher</h1>
 			</div>
 		);
